workreward/requirements.txt
asgiref==3.8.1
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.2
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
